Remove storefront from AppProxy. App Proxy is explicit enough
