+++ packages/shopify-app-remix/src/server/authenticate/public/types.ts
@@ -1,10 +1,10 @@
 import {ShopifyRestResources} from '@shopify/shopify-api';
 
 import type {AuthenticateCheckout} from './checkout/types';
-import type {AuthenticateStorefrontAppProxy} from './storefontAppProxy/types';
+import type {AuthenticateAppProxy} from './appProxy/types';
 
 export type AuthenticatePublic<Resources extends ShopifyRestResources> =
   AuthenticateCheckout & {
     checkout: AuthenticateCheckout;
-    storefrontAppProxy: AuthenticateStorefrontAppProxy<Resources>;
+    appProxy: AuthenticateAppProxy<Resources>;
   };
